SubImage + Scaled image rendering; documentation
The rendering API now allows for rendering sub-sections of an image,
e.g. for a SpriteSheet, as well as scaling of Images and SubImages.

I've documented the calls so the godocs thing should be a bit more
useful.

I've handled an error where if you try to load an image during the
init() of a package, it will error because the RenderInstance hasn't
been initialized. This now is handled and returned as an error message.